# services/identity/internal/api/contract/identity.gen.go
//go:build go1.22

// Package contract provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package contract

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for Action.
const (
	AssignComment        Action = "assign_comment"
	BootstrapPullRequest Action = "bootstrap_pull_request"
)

// Valid indicates whether the value is a known member of the Action enum.
func (e Action) Valid() bool {
	switch e {
	case AssignComment:
		return true
	case BootstrapPullRequest:
		return true
	default:
		return false
	}
}

// Defines values for ErrorCode.
const (
	BootstrapConflict   ErrorCode = "bootstrap_conflict"
	BootstrapIncomplete ErrorCode = "bootstrap_incomplete"
	BootstrapStale      ErrorCode = "bootstrap_stale"
	Forbidden           ErrorCode = "forbidden"
	ForgeryRejected     ErrorCode = "forgery_rejected"
	IdentityUnavailable ErrorCode = "identity_unavailable"
	InvalidRequest      ErrorCode = "invalid_request"
	MethodNotAllowed    ErrorCode = "method_not_allowed"
	NotFound            ErrorCode = "not_found"
	ReconnectRequired   ErrorCode = "reconnect_required"
)

// Valid indicates whether the value is a known member of the ErrorCode enum.
func (e ErrorCode) Valid() bool {
	switch e {
	case BootstrapConflict:
		return true
	case BootstrapIncomplete:
		return true
	case BootstrapStale:
		return true
	case Forbidden:
		return true
	case ForgeryRejected:
		return true
	case IdentityUnavailable:
		return true
	case InvalidRequest:
		return true
	case MethodNotAllowed:
		return true
	case NotFound:
		return true
	case ReconnectRequired:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatus.
const (
	Ok HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// Defines values for SessionState.
const (
	Authenticated SessionState = "authenticated"
	Disabled      SessionState = "disabled"
	SignedOut     SessionState = "signed_out"
)

// Valid indicates whether the value is a known member of the SessionState enum.
func (e SessionState) Valid() bool {
	switch e {
	case Authenticated:
		return true
	case Disabled:
		return true
	case SignedOut:
		return true
	default:
		return false
	}
}

// Action defines model for Action.
type Action string

// Authorization defines model for Authorization.
type Authorization struct {
	Action       Action    `json:"action"`
	CheckedAt    time.Time `json:"checked_at"`
	Repository   string    `json:"repository"`
	RepositoryId int64     `json:"repository_id"`
	User         User      `json:"user"`
}

// AuthorizationRequest defines model for AuthorizationRequest.
type AuthorizationRequest struct {
	Action Action `json:"action"`
	Csrf   string `json:"csrf"`
}

// BootstrapRequest defines model for BootstrapRequest.
type BootstrapRequest struct {
	BaseRevision string `json:"base_revision"`
	Csrf         string `json:"csrf"`
	Digest       string `json:"digest"`
}

// BootstrapResult defines model for BootstrapResult.
type BootstrapResult struct {
	Branch            string `json:"branch"`
	HeadSha           string `json:"head_sha"`
	PullRequestNumber int    `json:"pull_request_number"`
	PullRequestUrl    string `json:"pull_request_url"`
	Repository        string `json:"repository"`
}

// Error defines model for Error.
type Error struct {
	Code    ErrorCode `json:"code"`
	Message string    `json:"message"`
}

// ErrorCode defines model for Error.Code.
type ErrorCode string

// Health defines model for Health.
type Health struct {
	Status HealthStatus `json:"status"`
}

// HealthStatus defines model for Health.Status.
type HealthStatus string

// ProtectedRequest defines model for ProtectedRequest.
type ProtectedRequest struct {
	Csrf string `json:"csrf"`
}

// Session defines model for Session.
type Session struct {
	Csrf      *string      `json:"csrf"`
	ExpiresAt *time.Time   `json:"expires_at"`
	State     SessionState `json:"state"`
	User      *User        `json:"user"`
}

// SessionState defines model for Session.State.
type SessionState string

// SignOutResult defines model for SignOutResult.
type SignOutResult struct {
	Message string `json:"message"`
	Revoked bool   `json:"revoked"`
}

// User defines model for User.
type User struct {
	Id    int64  `json:"id"`
	Login string `json:"login"`
}

// Origin defines model for Origin.
type Origin = string

// AuthorizeRepositoryParams defines parameters for AuthorizeRepository.
type AuthorizeRepositoryParams struct {
	// Origin Exact configured public origin, required for every POST.
	Origin Origin `json:"Origin"`
}

// CreateBootstrapPullRequestParams defines parameters for CreateBootstrapPullRequest.
type CreateBootstrapPullRequestParams struct {
	// Origin Exact configured public origin, required for every POST.
	Origin Origin `json:"Origin"`
}

// StartSignInParams defines parameters for StartSignIn.
type StartSignInParams struct {
	// Origin Exact configured public origin, required for every POST.
	Origin Origin `json:"Origin"`
}

// CompleteSignInParams defines parameters for CompleteSignIn.
type CompleteSignInParams struct {
	State string `form:"state" json:"state"`
	Code  string `form:"code" json:"code"`
}

// SignOutParams defines parameters for SignOut.
type SignOutParams struct {
	// Origin Exact configured public origin, required for every POST.
	Origin Origin `json:"Origin"`
}

// AuthorizeRepositoryJSONRequestBody defines body for AuthorizeRepository for application/json ContentType.
type AuthorizeRepositoryJSONRequestBody = AuthorizationRequest

// CreateBootstrapPullRequestJSONRequestBody defines body for CreateBootstrapPullRequest for application/json ContentType.
type CreateBootstrapPullRequestJSONRequestBody = BootstrapRequest

// SignOutJSONRequestBody defines body for SignOut for application/json ContentType.
type SignOutJSONRequestBody = ProtectedRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)

	// (HEAD /health)
	HeadHealth(w http.ResponseWriter, r *http.Request)

	// (POST /v1/repositories/{owner}/{repo}/authorization)
	AuthorizeRepository(w http.ResponseWriter, r *http.Request, owner string, repo string, params AuthorizeRepositoryParams)

	// (POST /v1/repositories/{owner}/{repo}/bootstrap-pull-request)
	CreateBootstrapPullRequest(w http.ResponseWriter, r *http.Request, owner string, repo string, params CreateBootstrapPullRequestParams)

	// (GET /v1/session)
	GetSession(w http.ResponseWriter, r *http.Request)

	// (POST /v1/sign-in)
	StartSignIn(w http.ResponseWriter, r *http.Request, params StartSignInParams)

	// (GET /v1/sign-in/callback)
	CompleteSignIn(w http.ResponseWriter, r *http.Request, params CompleteSignInParams)

	// (POST /v1/sign-out)
	SignOut(w http.ResponseWriter, r *http.Request, params SignOutParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// HeadHealth operation middleware
func (siw *ServerInterfaceWrapper) HeadHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HeadHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AuthorizeRepository operation middleware
func (siw *ServerInterfaceWrapper) AuthorizeRepository(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "owner" -------------
	var owner string

	err = runtime.BindStyledParameterWithOptions("simple", "owner", r.PathValue("owner"), &owner, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "owner", Err: err})
		return
	}

	// ------------- Path parameter "repo" -------------
	var repo string

	err = runtime.BindStyledParameterWithOptions("simple", "repo", r.PathValue("repo"), &repo, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "repo", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params AuthorizeRepositoryParams

	headers := r.Header

	// ------------- Required header parameter "Origin" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Origin")]; found {
		var Origin Origin
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Origin", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Origin", valueList[0], &Origin, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Origin", Err: err})
			return
		}

		params.Origin = Origin

	} else {
		err := fmt.Errorf("Header parameter Origin is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Origin", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthorizeRepository(w, r, owner, repo, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateBootstrapPullRequest operation middleware
func (siw *ServerInterfaceWrapper) CreateBootstrapPullRequest(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "owner" -------------
	var owner string

	err = runtime.BindStyledParameterWithOptions("simple", "owner", r.PathValue("owner"), &owner, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "owner", Err: err})
		return
	}

	// ------------- Path parameter "repo" -------------
	var repo string

	err = runtime.BindStyledParameterWithOptions("simple", "repo", r.PathValue("repo"), &repo, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "repo", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateBootstrapPullRequestParams

	headers := r.Header

	// ------------- Required header parameter "Origin" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Origin")]; found {
		var Origin Origin
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Origin", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Origin", valueList[0], &Origin, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Origin", Err: err})
			return
		}

		params.Origin = Origin

	} else {
		err := fmt.Errorf("Header parameter Origin is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Origin", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateBootstrapPullRequest(w, r, owner, repo, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSession operation middleware
func (siw *ServerInterfaceWrapper) GetSession(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSession(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StartSignIn operation middleware
func (siw *ServerInterfaceWrapper) StartSignIn(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params StartSignInParams

	headers := r.Header

	// ------------- Required header parameter "Origin" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Origin")]; found {
		var Origin Origin
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Origin", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Origin", valueList[0], &Origin, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Origin", Err: err})
			return
		}

		params.Origin = Origin

	} else {
		err := fmt.Errorf("Header parameter Origin is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Origin", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StartSignIn(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CompleteSignIn operation middleware
func (siw *ServerInterfaceWrapper) CompleteSignIn(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params CompleteSignInParams

	// ------------- Required query parameter "state" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "state", r.URL.Query(), &params.State, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "state"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "state", Err: err})
		}
		return
	}

	// ------------- Required query parameter "code" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "code", r.URL.Query(), &params.Code, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "code"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "code", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CompleteSignIn(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SignOut operation middleware
func (siw *ServerInterfaceWrapper) SignOut(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params SignOutParams

	headers := r.Header

	// ------------- Required header parameter "Origin" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Origin")]; found {
		var Origin Origin
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Origin", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Origin", valueList[0], &Origin, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Origin", Err: err})
			return
		}

		params.Origin = Origin

	} else {
		err := fmt.Errorf("Header parameter Origin is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Origin", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SignOut(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/health", wrapper.GetHealth)
	m.HandleFunc(http.MethodHead+" "+options.BaseURL+"/health", wrapper.HeadHealth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/session", wrapper.GetSession)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/sign-in", wrapper.StartSignIn)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/sign-in/callback", wrapper.CompleteSignIn)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/sign-out", wrapper.SignOut)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/repositories/{owner}/{repo}/authorization", wrapper.AuthorizeRepository)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/repositories/{owner}/{repo}/bootstrap-pull-request", wrapper.CreateBootstrapPullRequest)

	return m
}